remove REACT_APP_MOCK
