--- packages/dashboard/src/mock/fake-authenticator.ts
@@ -1,20 +1,0 @@
-import EventEmitter from 'eventemitter3';
-import Authenticator, { AuthenticatorEventType } from '../components/auth/authenticator';
-import { User } from '../components/auth/user';
-
-export default class FakeAuthenticator extends EventEmitter<AuthenticatorEventType>
-  implements Authenticator {
-  readonly user: User = {
-    username: 'fakeUser',
-  };
-
-  async init(): Promise<void> {}
-
-  login(): Promise<never> {
-    throw new Error('Method not implemented.');
-  }
-
-  logout(): Promise<never> {
-    throw new Error('Method not implemented.');
-  }
-}
